Tests/LibWeb/Text/expected/css/FontFace-binary-data.txt

badFace.status: error
badFace.family: My Cool Font
zeroBytesFace.status: error
zeroBytesFace.family: Empty Font
face.status: loaded
face.family: Hash Sans
